refactor: move Janus to lib folder
