+++ lib/Janus.js
